Started reworking logger to support timestamps in log files
