+++ app/config/index.ts
@@ -1,5 +1,4 @@
 import app from './app'
 import constants from './constants'
-import logger from './logger'
 
-export { app, logger, constants }
+export { app, constants }
