+++ app/index.ts
@@ -1,5 +1,6 @@
 import { app, constants } from './config'
+import logger from './logger'
 
 app.listen(constants.PORT, () => {
-    console.log(`Application listening on port ${constants.PORT}.`)
+    logger.info(`Application listening on port ${constants.PORT}.`)
 })
